requirements.txt
fastapi[standard]
requests  
python-dotenv